+++ bluray_time.c
@@ -29,7 +29,7 @@ void bluray_duration_length(char *str, const uint64_t duration) {
 
 }
 
-uint64_t bluray_chapter_duration(struct bluray *bd, const uint32_t title_ix, const uint32_t chapter_ix, const uint32_t angle_ix) {
+uint64_t bluray_chapter_duration(struct bluray *bd, const uint32_t title_ix, const uint32_t chapter_ix, const uint8_t angle_ix) {
 
 	int retval = 0;
 	retval = bd_select_title(bd, title_ix);
@@ -61,7 +61,7 @@ uint64_t bluray_chapter_duration(struct bluray *bd, const uint32_t title_ix, con
 
 }
 
-void bluray_chapter_length(char *dest_str, struct bluray *bd, const uint32_t title_ix, const uint32_t chapter_ix, const uint32_t angle_ix) {
+void bluray_chapter_length(char *dest_str, struct bluray *bd, const uint32_t title_ix, const uint32_t chapter_ix, const uint8_t angle_ix) {
 
 	uint64_t duration = 0;
